memrchr: Fix alignment error in memrchr's fallback.
The fallback did not compute the offset correctly. It was intentioned to
land on usize-aligned addresses but did not. This resulted in a SIGBUS
on an ARMv7 platform!

I think like this, if we have a slice with pointer `ptr` and length
`len`, we want to find the last usize-aligned offset in the slice.

The correct computation should be:

For example if ptr = 1 and len = 6, and size_of::<usize>() is 4:

 [ x x x x x x ]
   1 2 3 4 5 6
         ^-- last aligned address at offset 3 from the start.

The last aligned address is ptr + len - (ptr + len) % usize_size.

Compute offset from the start as:

offset = len - (ptr + len) % usize_size = 6 - (1 + 6) % 4 = 6 - 3 = 3.

I believe the answer was always correct previously, if the platform
supported unaligned addresses.