tests: increase coverage dramatically
This commit updates the tests to increase their coverage dramatically.
They were influenced by my experimenting with the existing tests by
introducing subtle bugs into the memchr implementations, and witnessing
with horror at the test suite continuing to pass.

We increase by coverage by doing two things. Firstly, we pad the size of
the corpus of each test by adding bytes that we never use as a needle.
We go all the way up to ~500 bytes to make sure we cover all forms of
case analysis. We pad both sides of the corpus and update our expected
positions accordingly.

Secondly, for each `find` test, we vary the alignment of the slice we
pass to memchr, all the way up to 66 to account for AVX2 optimizations
(which will read at least 32 bytes at a time, but loop unrolling might
make this 64 bytes at a time).

This coverage passed my previous smell test. Every bug I tried to
introduce is now caught by the test suite (unless the bug introduces UB,
but that's par for the course).