fallback: fix unaligned read
This commit fixes a bug that caused undefined behavior by doing an
aligned read when an unaligned read was intended. In particular, we were
casting a `*const u8` to a `*const usize` and then reading its contents
to a `*mut usize`. The correct implementation is to reverse this such
that we read directly from `*const u8` and write to `*mut u8`, such that
ptr::copy_nonoverlapping knows to emit an unaligned load from the
`*const u8` pointer.