Use static with AtomicPtr instead of static mut for ifunc
There's a faint rumbling in the distance, that static mut is to be
avoided. It seems to be a problem when we take *mutable* references to
the static, and this code didn't.

Still, port the code to using a regular static with an atomic. We can
fortunately use a fn to raw pointer cast (fn to usize is not available
in statics).