ifunc: use static with AtomicPtr
This replaces the use of `static mut`, because there's a faint rumbling
in the distance that static mut is to be avoided[1]. It seems to be a
problem when we take *mutable* references to the static, and this code
didn't. Therefore, this is only done out of an abundance of caution;
there is no belief that the previous code was incorrect.

Still, we port the code to using a regular static with an atomic.
We can fortunately use a fn to raw pointer cast (fn to usize is not
available in statics).

Closes #42

[1] - https://github.com/rust-lang/rust/issues/53639