simd: Remove redundant block in memchr3
In the sse2 and avx implementation of memchr3.

The memchr3 steps by two vector-sizes at time. When a match was found in
one of them, it would check the first one first, then the second one and
then the second one again.

It wasn't causing any incorrect behavior, but did extra work unless
compiler optimized it away.