cfg: support MIRI by using `naive` implementations
MIRI has problems dealing with SIMD or raw pointers, so we cannot use
the fallback or SIMD accelerated versions. Instead, we fall back to the
"naive" version using iterators.

Ideally, this is a temporary measure to make more things work with MIRI
in practice. In the future, this work-around shouldn't be needed.

Closes #51