bench: add another run
This is specifically meant to capture timings from bstr and regex after
causing them to use the memmem implementation. It's a sanity check to
make sure everything looks okay. Indeed, particularly with respect to
regex, the only difference appears to be a bit worse latency. Fixing the
latency of regexes is a hill to climb for another day.