Port the memchr implementation to "genericsimd"
This commit ports the various memchr algorithms implemented for SSE2 at
`src/memchr/x86/sse2.rs` to the `Vector` trait. This added a few minor
methods to the `Vector` trait and the rewrite is hopefully quite
straightforward.

Afterwards the `src/memchr/x86/avx.rs` module is deleted because the
theory is that it's simply the AVX2-using version of the new generic
code (basically a new type parameter). Further changes were made to all
the `memchr`-related functions to also enable WebAssembly acceleration
with the `v128` type. Ideally there's only one place necessary for new
architectures to specify their implementation as well.

This method of abstraction is distinct from the `memmem` module,
however. I've tried to draw inspiration from the previous design of the
`memchr` module but I believe that the `memmem` module is newer and so
further refactorings may be necessary to align the two in style.