Add FindIter::into_owned and FindRevIter::into_owned
`memmem::Finder::new(needle).into_owned().find_iter(haystack)` returns a
`FindIter` that borrows the `Finder`. That `Finder` owns a copy of the
needle.

`memmem::find_iter(haystack, needle)` returns a `FindIter` that owns a
`Finder` that borrows the needle.

There is currently no way to get a `FindIter` (or `FindRevIter`) that owns
a `Finder` that owns a copy of the needle. This prevents storing a `FindIter`
in some struct without `Box`ing the `Finder`.

This change adds just that.