bench: add WASM benchmark output
This is a long overdue benchmark of the WASM implementation that landed
in #84. I tried to do this a few months ago, but I mucked it up and
never got back around to it. (This is also why I haven't put out a new
release with the WASM implementation, since I wanted to do a sanity
benchmark run first.)