vim: add coc-rust-analyzer specific config
This specifically disables 'allFeatures' in the coc-rust-analyzer
config. This crate can't really be built with --all-features because of
its weird "dep of rust std" configuration.

It is perhaps a little weird to commit this dev environment specific
configuration to the repo, but this avoids me needing to copy this
project specific config to every checkout of memchr.

Big thanks to Hazel for teaching me about project specific overrides for
CoC: https://twitter.com/burntsushi5/status/1522560334267092995