add branchless finalizer to sse2 memchr
This should be extended to other contexts if others are able to observe
the same gains I was able to observe locally. See comment changes for an
explanation of what's going on here, but basically we can avoid some
looping if we eat an initial extra branch on whether our length is
greater than our loop size. We can apply a similar optimization to the
AVX2 case, and to memchr2 and friends.