benchmarks: update memchr version
Previously, we pinned memchr *and* pointed it to the current source
code. That means `rebar measure` would automatically benchmark changes.
But once a new release is out, CI fails because the current version has
been updated.

This is annoying. We need to figure out a better solution here. For now,
we require folks to uncomment the `patch` block to benchmark the current
code instead of what's on crates.io. Sigh.