msrv: bump to Rust 1.61
It was a gaffe to release memchr 2.6 with a Rust 1.60 MSRV. I had
thought we were good there, but it turns out I never actually tested it
on aarch64. On aarch64, while many of the intrinsics were stabilized by
Rust 1.60, the actual `#[target_feature(enable = "neon")]` attribute was
not enabled until Rust 1.61.

Since Rust 1.61 is just 6 weeks from Rust 1.60 and Rust 1.61 is still
pretty conservative overall (released almost over 1.5 years ago), we
just bump the MSRV to Rust 1.61 and declare the original Rust 1.60 a
bug.

Fixes #136