arch: add big endian regression test
This was reported along with the fact that memchr 2.7.3 broke big endian
targets. While I at first thought the test suite just somehow missed the
regression and thus this test was somehow novel in some interesting way,
it turned out it was just a CI configuration failure. The existing tests
catch the regression easily (dozens of test failures). The problem, as
mentioned in the previous commit, was that tests in CI weren't using
`cross` when cross compilation was enabled.

So while this test is seemingly redundant, I still think it's good form
to add it.

Ref https://github.com/BurntSushi/memchr/issues/152